more refactoring of the current sensing
